Separate stateful report from stateless one
