+++ packages/report-list/components/Report/story.jsx
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { checkA11y } from 'storybook-addon-a11y';
 import { action } from '@storybook/addon-actions';
 import Report from './index';
+import StatelessReport from './StatelessReport';
 
 
 const report = {
@@ -17,6 +18,14 @@ storiesOf('Report')
     <Report
       {...report}
       selectReport={action('select report')}
+    />
+  ))
+  .add('can display buttons for viewing and removing the report', () => (
+    <StatelessReport
+      {...report}
+      showButtons
+      selectReport={action('select report')}
+      removeReport={action('remove report')}
     />
   ));
